adds a hacky egl context, needs refactor, now getting gl output

+++ app/src/main/java/app/jerboa/gpgpu/data/GPUData.kt
@@ -4,5 +4,5 @@ data class GPUData(
     val time: Long,
     val textureTime: Long,
     val drawTime: Long,
-    val error: Long
+    val error: Float
 )
